Refactor integration test helpers
This commit performs three operations:

1. Move `integration-tests/helpers.js` to
   `integration-tests/helpers/index.js`.
2. Move `FakeAgent` from inside `integration-tests/helpers/index.js` to
   its own file called `integration-tests/helpers/fake_agent.js`.
3. Move the setting up of the Express app inside the `FakeAgent.start`
   method to a private helper function.
